Python URL generator script - BETTER THAN JAVA
This Python implementation will generate download URLs based on the m3u8 file and so, it is precise. It will generate the exact URLs. The Java implementation is like a trial and error method, which blindly generates URLs by iterating over a number limit. Unlike that, this Python implementation is precise and more user friendly.